+++ HandballShop/src/test/java/isa/handballshop/services/util/AccountingTest.java
@@ -10,13 +10,13 @@ public class AccountingTest{
     @Test
     public void testCalcoloPrezzo(){
         ArrayList<Double> prezzo = new ArrayList<Double>();
-        ArrayList<Integer> quantita = new ArrayList<Integer>();
+        ArrayList<Long> quantita = new ArrayList<Long>();
 
         prezzo.add((double) 3);
         prezzo.add((double) 5);
 
-        quantita.add(2);
-        quantita.add(4);
+        quantita.add((long) 2);
+        quantita.add((long) 4);
 
         double result = Accounting.calcoloPrezzo(prezzo, quantita);
         assertEquals(26, result);
